requirements.txt
fastapi
uvicorn
supabase
langchain
sentence-transformers
faiss-cpu
transformers
requests
python-dotenv
streamlit